created DisplayReview component and organized into review folder
