--- tenttalk-ui/src/components/campground/DisplayReview.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function DisplayReview() {
-  return (
-    <div>DisplayReview</div>
-  )
-}
