+++ tenttalk-ui/src/components/review/DisplayReview.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function DisplayReview({reviewId}) {
+
+
+  
+  return (
+    <div>
+      
+    </div>
+  )
+}
